add mara-pipelines click command group
